refactor(api-bundle): address review feedback on token exchange
- move the programmatic-token exchange from StorageApiTokenAuthenticator
  into StorageApiTokenFactory, so both token-creation paths (legacy
  verification and exchange) live in the factory and the authenticator
  only extracts, routes and authorizes
- make the resolver client and logger mandatory dependencies (no
  use-case without them)
- drop the hardcoded ServiceDnsType::INTERNAL for the resolver client;
  it now follows the ServiceClient's configured default DNS type like
  every other ManageApiClientFactory method
