+++ libs/api-bundle/tests/Security/StorageApiToken/StorageApiTokenFactoryTest.php
@@ -6,7 +6,12 @@ namespace Keboola\ApiBundle\Tests\Security\StorageApiToken;
 
 use Exception;
 use Generator;
+use GuzzleHttp\Exception\ConnectException;
+use GuzzleHttp\Psr7\Request as GuzzleRequest;
 use Keboola\ApiBundle\Security\StorageApiToken\StorageApiTokenFactory;
+use Keboola\ManageApi\Client as ManageApiClient;
+use Keboola\ManageApi\ClientException as ManageApiClientException;
+use Keboola\ManageApi\MaintenanceException as ManageApiMaintenanceException;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
 use Keboola\StorageApi\MaintenanceException;
@@ -14,11 +19,38 @@ use Keboola\StorageApiBranch\ClientWrapper;
 use Keboola\StorageApiBranch\Factory\StorageClientRequestFactory;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use RuntimeException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationException;
 
 class StorageApiTokenFactoryTest extends TestCase
 {
+    private const SUBJECT_TOKEN = 'kbc_at_secret';
+    private const PROJECT_ID_HEADER = 'X-KBC-ProjectId';
+
+    private function createFactory(
+        ?StorageClientRequestFactory $clientRequestFactory = null,
+        ?ManageApiClient $resolverClient = null,
+        ?LoggerInterface $logger = null,
+    ): StorageApiTokenFactory {
+        return new StorageApiTokenFactory(
+            $clientRequestFactory ?? $this->createMock(StorageClientRequestFactory::class),
+            $resolverClient ?? $this->createMock(ManageApiClient::class),
+            $logger ?? new NullLogger(),
+        );
+    }
+
+    private function createProgrammaticTokenRequest(): Request
+    {
+        $request = Request::create('https://keboola.com');
+        $request->headers->set('Authorization', 'Bearer ' . self::SUBJECT_TOKEN);
+        $request->headers->set(self::PROJECT_ID_HEADER, '123');
+
+        return $request;
+    }
+
     // ---------------------------------------------------------------------------
     // createFromRequest – happy path
     // ---------------------------------------------------------------------------
@@ -50,7 +82,7 @@ class StorageApiTokenFactoryTest extends TestCase
         $request = Request::create('https://keboola.com');
         $request->headers->set(StorageClientRequestFactory::TOKEN_HEADER, 'tok');
 
-        $token = (new StorageApiTokenFactory($factoryMock))->createFromRequest($request);
+        $token = $this->createFactory(clientRequestFactory: $factoryMock)->createFromRequest($request);
 
         self::assertSame('tok', $token->getTokenValue());
     }
@@ -94,7 +126,7 @@ class StorageApiTokenFactoryTest extends TestCase
         self::expectExceptionMessage($expectedExceptionMessage);
         self::expectExceptionCode($expectedExceptionCode);
 
-        (new StorageApiTokenFactory($factoryMock))->createFromRequest($request);
+        $this->createFactory(clientRequestFactory: $factoryMock)->createFromRequest($request);
     }
 
     public static function provideExceptionData(): Generator
@@ -120,13 +152,26 @@ class StorageApiTokenFactoryTest extends TestCase
     }
 
     // ---------------------------------------------------------------------------
-    // createFromResolvedToken – request mutation and isolation
+    // createFromProgrammaticToken – exchange success, request mutation and isolation
     // ---------------------------------------------------------------------------
 
-    public function testCreateFromResolvedTokenPassesLegacyTokenAndDropsAuthorization(): void
+    public function testCreateFromProgrammaticTokenPassesLegacyTokenAndDropsAuthorization(): void
     {
         /** @var Request|null $capturedRequest */
         $capturedRequest = null;
+
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->with(123, self::SUBJECT_TOKEN)
+            ->willReturn([
+                'storageToken' => 'legacy-token',
+                'projectId' => 123,
+                'tokenId' => '42',
+                'userId' => '7',
+                'expiresAt' => null,
+            ]);
 
         $clientMock = $this->createMock(Client::class);
         $clientMock
@@ -153,12 +198,11 @@ class StorageApiTokenFactoryTest extends TestCase
                 return $wrapperMock;
             });
 
-        $originalRequest = Request::create('https://keboola.com');
-        $originalRequest->headers->set('Authorization', 'Bearer kbc_at_secret');
+        $originalRequest = $this->createProgrammaticTokenRequest();
         $originalRequest->headers->set('X-KBC-RunId', 'run-1');
 
-        $token = (new StorageApiTokenFactory($factoryMock))
-            ->createFromResolvedToken($originalRequest, 'legacy-token');
+        $token = $this->createFactory(clientRequestFactory: $factoryMock, resolverClient: $resolverClient)
+            ->createFromProgrammaticToken($originalRequest, self::SUBJECT_TOKEN);
 
         self::assertSame('legacy-token', $token->getTokenValue());
         self::assertNotNull($capturedRequest);
@@ -183,9 +227,301 @@ class StorageApiTokenFactoryTest extends TestCase
 
         // The ORIGINAL request must remain untouched.
         self::assertSame(
-            'Bearer kbc_at_secret',
+            'Bearer ' . self::SUBJECT_TOKEN,
             $originalRequest->headers->get('Authorization'),
             'Original request Authorization header must not be mutated.',
         );
+    }
+
+    // ---------------------------------------------------------------------------
+    // createFromProgrammaticToken – project id header validation
+    // ---------------------------------------------------------------------------
+
+    #[DataProvider('provideInvalidProjectIdHeaders')]
+    public function testCreateFromProgrammaticTokenThrowsWith400ForInvalidProjectIdHeader(?string $headerValue): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::never())
+            ->method('resolveStorageToken');
+
+        $request = Request::create('https://keboola.com');
+        $request->headers->set('Authorization', 'Bearer ' . self::SUBJECT_TOKEN);
+        if ($headerValue !== null) {
+            $request->headers->set(self::PROJECT_ID_HEADER, $headerValue);
+        }
+
+        $factory = $this->createFactory(resolverClient: $resolverClient);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken($request, self::SUBJECT_TOKEN),
+        );
+
+        self::assertSame(400, $exception->getCode());
+    }
+
+    public static function provideInvalidProjectIdHeaders(): Generator
+    {
+        yield 'missing header' => ['headerValue' => null];
+        yield 'empty string' => ['headerValue' => ''];
+        yield 'non-numeric string' => ['headerValue' => 'abc'];
+        yield 'zero' => ['headerValue' => '0'];
+        yield 'negative number' => ['headerValue' => '-5'];
+        // Past PHP_INT_MAX: ctype_digit would have accepted this and the cast would have silently
+        // wrapped to a different project id; filter_var rejects it.
+        yield 'overflows int range' => ['headerValue' => '99999999999999999999999999'];
+        yield 'leading zeros' => ['headerValue' => '0123'];
+    }
+
+    // ---------------------------------------------------------------------------
+    // createFromProgrammaticToken – resolver error mapping
+    // ---------------------------------------------------------------------------
+
+    #[DataProvider('provideResolverClientErrorMapping')]
+    public function testExchangeMapsResolverClientErrorToExpectedCode(int $statusCode, int $expectedCode): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willThrowException(new ManageApiClientException(self::SUBJECT_TOKEN, $statusCode));
+
+        $factory = $this->createFactory(resolverClient: $resolverClient);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame($expectedCode, $exception->getCode());
+        // The Manage API client embeds the response body in the exception message; assert our
+        // mapping never echoes it back so subject/storage token material cannot leak.
+        self::assertStringNotContainsString(self::SUBJECT_TOKEN, $exception->getMessageKey());
+    }
+
+    public static function provideResolverClientErrorMapping(): Generator
+    {
+        yield 'bad request -> 400' => ['statusCode' => 400, 'expectedCode' => 400];
+        yield 'unauthorized -> 401' => ['statusCode' => 401, 'expectedCode' => 401];
+        yield 'forbidden -> 403' => ['statusCode' => 403, 'expectedCode' => 403];
+        yield 'server error -> 502' => ['statusCode' => 500, 'expectedCode' => 502];
+        yield 'connect error without response (code 0) -> 502' => ['statusCode' => 0, 'expectedCode' => 502];
+        yield 'unexpected client error -> 502' => ['statusCode' => 418, 'expectedCode' => 502];
+    }
+
+    public function testExchangeMapsMaintenanceExceptionTo503(): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willThrowException(new ManageApiMaintenanceException('Maintenance', 30, []));
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects(self::once())
+            ->method('warning')
+            ->with(self::stringContains('maintenance'), ['projectId' => 123, 'retryAfter' => 30]);
+
+        $factory = $this->createFactory(resolverClient: $resolverClient, logger: $logger);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(503, $exception->getCode());
+    }
+
+    public function testExchangeMapsConnectExceptionTo502(): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willThrowException(new ConnectException('Connection refused', new GuzzleRequest('POST', 'resolve')));
+
+        $factory = $this->createFactory(resolverClient: $resolverClient);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(502, $exception->getCode());
+    }
+
+    public function testExchangeMapsServiceAccountTokenFailureTo502(): void
+    {
+        // KubernetesServiceAccountTokenAuthenticationStrategy throws RuntimeException when the
+        // projected SA token file is missing/unreadable/empty.
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willThrowException(new RuntimeException('token file is empty'));
+
+        $factory = $this->createFactory(resolverClient: $resolverClient);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(502, $exception->getCode());
+    }
+
+    #[DataProvider('provideInvalidResolverResponses')]
+    public function testExchangeThrowsWith502WhenResolverResponseLacksStorageToken(mixed $response): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willReturn($response);
+
+        $clientRequestFactory = $this->createMock(StorageClientRequestFactory::class);
+        $clientRequestFactory
+            ->expects(self::never())
+            ->method('createClientWrapper');
+
+        $factory = $this->createFactory(
+            clientRequestFactory: $clientRequestFactory,
+            resolverClient: $resolverClient,
+        );
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(502, $exception->getCode());
+    }
+
+    public static function provideInvalidResolverResponses(): Generator
+    {
+        yield 'missing storageToken' => ['response' => ['projectId' => 123]];
+        yield 'empty storageToken' => ['response' => ['storageToken' => '']];
+        yield 'non-string storageToken' => ['response' => ['storageToken' => 123]];
+    }
+
+    // ---------------------------------------------------------------------------
+    // createFromProgrammaticToken – failure logging
+    // ---------------------------------------------------------------------------
+
+    public function testExchangeLogsUnexpectedResolverStatusWithoutTokenMaterial(): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            // The Manage client embeds the response body (here the subject token) in the message.
+            ->willThrowException(new ManageApiClientException(self::SUBJECT_TOKEN, 500));
+
+        $loggedContext = null;
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects(self::once())
+            ->method('warning')
+            ->willReturnCallback(function (string $message, array $context) use (&$loggedContext): void {
+                self::assertStringNotContainsString(self::SUBJECT_TOKEN, $message);
+                $loggedContext = $context;
+            });
+
+        $factory = $this->createFactory(resolverClient: $resolverClient, logger: $logger);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(502, $exception->getCode());
+        self::assertSame(['projectId' => 123, 'resolverStatus' => 500], $loggedContext);
+        // Defense in depth: the subject token must never reach the log context.
+        self::assertStringNotContainsString(self::SUBJECT_TOKEN, json_encode($loggedContext, JSON_THROW_ON_ERROR));
+    }
+
+    public function testExchangeLogsMissingStorageTokenWithoutLoggingResolverResponse(): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willReturn(['storageToken' => '', 'leaked' => 'super-secret-storage-token']);
+
+        $loggedContext = null;
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects(self::once())
+            ->method('warning')
+            ->willReturnCallback(function (string $message, array $context) use (&$loggedContext): void {
+                $loggedContext = $context;
+            });
+
+        $factory = $this->createFactory(resolverClient: $resolverClient, logger: $logger);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(502, $exception->getCode());
+        self::assertSame(['projectId' => 123], $loggedContext);
+        // The raw resolver response (which may carry a storageToken value) must not be logged.
+        self::assertStringNotContainsString(
+            'super-secret-storage-token',
+            json_encode($loggedContext, JSON_THROW_ON_ERROR),
+        );
+    }
+
+    public function testExchangeDoesNotLogWarningForClientFaultStatuses(): void
+    {
+        $resolverClient = $this->createMock(ManageApiClient::class);
+        $resolverClient
+            ->expects(self::once())
+            ->method('resolveStorageToken')
+            ->willThrowException(new ManageApiClientException('unauthorized', 401));
+
+        // 401/403/400 are client faults, not our-side incidents, so they must not raise a warning.
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects(self::never())
+            ->method('warning');
+
+        $factory = $this->createFactory(resolverClient: $resolverClient, logger: $logger);
+
+        $exception = $this->captureAuthException(
+            fn() => $factory->createFromProgrammaticToken(
+                $this->createProgrammaticTokenRequest(),
+                self::SUBJECT_TOKEN,
+            ),
+        );
+
+        self::assertSame(401, $exception->getCode());
+    }
+
+    private function captureAuthException(callable $fn): CustomUserMessageAuthenticationException
+    {
+        try {
+            $fn();
+        } catch (CustomUserMessageAuthenticationException $e) {
+            return $e;
+        }
+
+        self::fail('Expected CustomUserMessageAuthenticationException was not thrown.');
     }
 }
